# public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 'API macarte'
  version: '1.0'
servers:
  -
    url: 'https://macarte-qualif.ign.fr/'
    description: 'Accède aux données de l''environnement Ma Carte'
paths:
...
  '/api/users/public/{public_id}':
    get:
      tags:
        - User
      parameters:
        -
          name: publicname
          in: path
          description: 'Identifiant public (immuable) de l''utilisateur'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Attributs publics de l''utilisateur'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_public'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
...
    User_public:
      properties:
        public_id:
          description: 'identifiant public de l''utilisateur, ne sera jamais modifié'
          type: string
        public_name:
          description: 'Unique, ne doit pas être déjà utilisé (Erreur 400)'
          type: string
        twitter_account:
          description: 'Nom de compte Twitter'
          type: string
        facebook_account:
          description: 'Nom de compte Facebook'
          type: string
        linkedin_account:
          description: 'Nom de compte LinkedIn'
          type: string
        presentation:
          description: 'Présentation publique'
          type: string
        profile_picture:
          description: 'Url de l''image de profil'
          type: string
          format: uri
        cover_picture:
          description: 'Url de l''image de couverture'
          type: string
          format: uri
      type: object
...
  responses:
...
    NotFound:
      description: 'La resource n''existe pas'
      content:
        application/json:
          schema:
            properties:
              code:
                type: integer
                example: '404'
              message:
                type: string
                example: 'La resource n''existe pas'
            type: object
...
  securitySchemes:
    bearer:
      type: apiKey
      name: Authorization
      in: header
      bearerFormat: JWT
      scheme: Bearer